% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obis.R
\name{obis_taxon}
\alias{obis_taxon}
\title{Look up OBIS taxon by ID}
\usage{
obis_taxon(taxon_id)
}
\arguments{
\item{taxon_id}{OBIS taxon ID (integer)}
}
\value{
tibble: taxonID, scientificName, taxonomicStatus, taxonRank,
kingdom, phylum, class, order, family, genus
}
\description{
Look up OBIS taxon by ID
}
